fog overlay when in hotzone
New fog effect canvas (could be useful more generally), in which a child object with orange fog image is activated when Camera.current collides with the HotZone.

For some reason this activates even when the third person camera is the active one, so will take a look at this another time. But works well enough for first person at the moment.